backend/api-gateway/requirements.txt
requests
Flask==2.3.2
Flask-Limiter==2.7.0
Flask-CORS
Flask-Session
redis
merchant-integration-service-0.1.1.tar.gz
loan-service-0.1.1.tar.gz
credit-service-0.1.1.tar.gz

grpcio
grpcio-tools